% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console.R
\name{.eri_say_done}
\alias{.eri_say_done}
\title{Emit a success confirmation, only at "full" verbosity.}
\usage{
.eri_say_done(..., .envir = parent.frame())
}
\description{
Emit a success confirmation, only at "full" verbosity.
}
\keyword{internal}
