% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console.R
\name{.eri_say_info}
\alias{.eri_say_info}
\title{Emit an informational line, only at "full" verbosity.}
\usage{
.eri_say_info(..., .envir = parent.frame())
}
\description{
Emit an informational line, only at "full" verbosity.
}
\keyword{internal}
