% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odk.R
\name{.odk_creds}
\alias{.odk_creds}
\title{Resolve url/auth from an odk_connection or explicit args}
\usage{
.odk_creds(con, url, auth)
}
\description{
Resolve url/auth from an odk_connection or explicit args
}
\keyword{internal}
