% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dq.R
\name{dq_result-methods}
\alias{dq_result-methods}
\alias{print.dq_result}
\alias{summary.dq_result}
\title{S3 methods for dq_result objects}
\usage{
\method{print}{dq_result}(x, ...)

\method{summary}{dq_result}(object, ...)
}
\arguments{
\item{x, object}{A \code{dq_result} object returned by \code{run_dq_checks()}.}

\item{...}{Unused; included for S3 method compatibility.}
}
\description{
S3 methods for dq_result objects
}
