% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_lf.R
\name{eri_lf_program_levels}
\alias{eri_lf_program_levels}
\title{Standard LF programme status levels}
\usage{
eri_lf_program_levels()
}
\value{
A character vector of length 5.
}
\description{
Returns the canonical ordered character vector of WHO/GPELF LF elimination
status levels. Use with \code{factor(status_col, levels = eri_lf_program_levels())}
to ensure correct ordering in tables and maps.
}
\examples{
\dontrun{
data |>
  dplyr::mutate(status = factor(status, levels = eri_lf_program_levels()))
}
}
