% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_oncho.R
\name{eri_oncho_program_levels}
\alias{eri_oncho_program_levels}
\title{OEPA onchocerciasis program status levels}
\usage{
eri_oncho_program_levels()
}
\value{
Character vector of length 5.
}
\description{
Returns the ordered vector of OEPA oncho program status levels used for
plotting and data validation. Levels are ordered from no endemicity to
verified elimination.
}
\examples{
eri_oncho_program_levels()
}
