% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports_pptx.R
\name{eri_pptx_add_section}
\alias{eri_pptx_add_section}
\title{Add a section divider slide to an ERI PowerPoint}
\usage{
eri_pptx_add_section(pptx, title)
}
\arguments{
\item{pptx}{An \code{officer} rpptx object.}

\item{title}{Character; section title text.}
}
\value{
The updated rpptx object.
}
\description{
Add a section divider slide to an ERI PowerPoint
}
