% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports_pptx.R
\name{eri_pptx_save}
\alias{eri_pptx_save}
\title{Save an ERI PowerPoint to disk}
\usage{
eri_pptx_save(pptx, path)
}
\arguments{
\item{pptx}{An \code{officer} rpptx object.}

\item{path}{Character; output file path (should end in \code{.pptx}).}
}
\value{
\code{path} invisibly.
}
\description{
Save an ERI PowerPoint to disk
}
\examples{
\dontrun{
eri_pptx_save(pptx, "outputs/malaria_report.pptx")
}
}
