% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odk.R
\name{list_odk_app_users}
\alias{list_odk_app_users}
\title{List all users in ODK}
\usage{
list_odk_app_users(
  url = Sys.getenv("ODK_URL"),
  auth = Sys.getenv("ODK_TOKEN"),
  project_id,
  testing = FALSE
)
}
\arguments{
\item{url}{\code{chr} Target URL for the ODK Central API}

\item{auth}{\code{chr} Authorization token to access URL}

\item{project_id}{\code{int} The project id for which you want to identify users}

\item{testing}{\code{bool} T/F if you want to just verify that the API if working}
}
\value{
\code{tibble} Output of all projects from API call
}
\description{
Given verified access list all users
}
