% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odk.R
\name{list_odk_form_users}
\alias{list_odk_form_users}
\title{List users assigned to an ODK form}
\usage{
list_odk_form_users(
  con = NULL,
  url = Sys.getenv("ODK_URL"),
  auth = Sys.getenv("ODK_TOKEN"),
  project_id,
  form_id
)
}
\arguments{
\item{con}{An \code{odk_connection} from \code{\link[=init_odk_connection]{init_odk_connection()}}, or \code{NULL} to use env vars}

\item{url}{\code{chr} Server URL (used when \code{con = NULL})}

\item{auth}{\code{chr} Bearer token (used when \code{con = NULL})}

\item{project_id}{\code{int} Project ID from \code{\link[=list_odk_projects]{list_odk_projects()}}}

\item{form_id}{\code{chr} Form ID from \code{\link[=list_odk_forms]{list_odk_forms()}}}
}
\value{
\code{tibble} of assigned users and roles
}
\description{
List users assigned to an ODK form
}
