% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odk.R
\name{list_odk_projects}
\alias{list_odk_projects}
\title{List ODK projects}
\usage{
list_odk_projects(
  con = NULL,
  url = Sys.getenv("ODK_URL"),
  auth = Sys.getenv("ODK_TOKEN")
)
}
\arguments{
\item{con}{An \code{odk_connection} from \code{\link[=init_odk_connection]{init_odk_connection()}}, or \code{NULL} to use env vars}

\item{url}{\code{chr} Server URL (used when \code{con = NULL})}

\item{auth}{\code{chr} Bearer token (used when \code{con = NULL})}
}
\value{
\code{tibble} with columns \code{project_id}, \code{project}, \code{description}
}
\description{
List ODK projects
}
