% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odk.R
\name{download_form_attachments}
\alias{download_form_attachments}
\title{See all users who have access to a form}
\usage{
download_form_attachments(
  url = Sys.getenv("ODK_URL"),
  auth = Sys.getenv("ODK_TOKEN"),
  project_id,
  form_id,
  folder_loc,
  image_label,
  other_vars,
  add_condition = FALSE,
  condition = NULL
)
}
\arguments{
\item{url}{\code{chr} Target URL for the ODK Central API}

\item{auth}{\code{chr} Authorization token to access URL}

\item{project_id}{\code{int} The project id for which you want to identify attachments}

\item{form_id}{\code{chr} The form id for which we want to identify attachments}

\item{folder_loc}{\code{chr} The folder in which all images should be dumped out}

\item{image_label}{\code{chr} The variable name that will be used for the image label}

\item{other_vars}{\code{chr} A concatenated character vector of other variables to include}

\item{add_condition}{\code{bool} Determine if a condition should be used}

\item{condition}{\code{chr} The dplyr::filter() command to filter the downloaded data}
}
\value{
\code{tibble} Output of all attachment names and their links to the forms
}
\description{
Pull all media attachments from a form
}
