% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odk.R
\name{download_odk_form}
\alias{download_odk_form}
\title{Download ODK form data}
\usage{
download_odk_form(
  url = Sys.getenv("ODK_URL"),
  auth = Sys.getenv("ODK_TOKEN"),
  project_id,
  form_id,
  testing = FALSE,
  attachments = FALSE
)
}
\arguments{
\item{url}{\code{chr} Target URL for the ODK Central API}

\item{auth}{\code{chr} Authorization token to access URL}

\item{project_id}{\code{int} Project id from \code{list_odk_projects()}}

\item{form_id}{\code{chr} From id from \code{list_odk_forms()}}

\item{testing}{\code{bool} T/F if you want to just verify that the API if working}

\item{attachments}{\code{bool} T/F if you want to include attachments from the form}
}
\value{
\code{tibble} Download of all data from an ODK form
}
\description{
Given verified access, project id and form ID, download all data from an
ODK form
}
