% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{eri_delete}
\alias{eri_delete}
\title{Delete a file}
\usage{
eri_delete(file_loc, azure = TRUE, azcontainer = NULL)
}
\arguments{
\item{file_loc}{\code{str} Path of file.}

\item{azure}{\code{logical} Whether the function should interact with the TCC Azure environment.
Defaults to \code{TRUE}, otherwise, interacts with files locally.}

\item{azcontainer}{\verb{Azure container} A container object returned by
\code{\link[=get_azure_storage_connection]{get_azure_storage_connection()}}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Thin wrapper around \code{\link[=erifunctions_io]{erifunctions_io()}} for deleting a file.
}
