% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{eri_list}
\alias{eri_list}
\title{List files in a directory}
\usage{
eri_list(file_loc = "", full_names = TRUE, azure = TRUE, azcontainer = NULL)
}
\arguments{
\item{file_loc}{\code{str} Path of file.}

\item{full_names}{\code{logical} If \code{io="list"}, include the full reference path. Default \code{TRUE}.}

\item{azure}{\code{logical} Whether the function should interact with the TCC Azure environment.
Defaults to \code{TRUE}, otherwise, interacts with files locally.}

\item{azcontainer}{\verb{Azure container} A container object returned by
\code{\link[=get_azure_storage_connection]{get_azure_storage_connection()}}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Thin wrapper around \code{\link[=erifunctions_io]{erifunctions_io()}} for listing directory contents.
}
