% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{eri_write}
\alias{eri_write}
\title{Write an object to a file}
\usage{
eri_write(obj, file_loc, ..., azure = TRUE, azcontainer = NULL)
}
\arguments{
\item{obj}{\code{str} Object to be loaded into Azure}

\item{file_loc}{\code{str} Path of file.}

\item{...}{Optional parameters that work with \code{\link[readr:read_delim]{readr::read_delim()}} or \code{\link[readxl:read_excel]{readxl::read_excel()}}.}

\item{azure}{\code{logical} Whether the function should interact with the TCC Azure environment.
Defaults to \code{TRUE}, otherwise, interacts with files locally.}

\item{azcontainer}{\verb{Azure container} A container object returned by
\code{\link[=get_azure_storage_connection]{get_azure_storage_connection()}}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Thin wrapper around \code{\link[=erifunctions_io]{erifunctions_io()}} for writing files.
}
