% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{get_azure_storage_connection}
\alias{get_azure_storage_connection}
\title{Validate connection to Azure}
\usage{
get_azure_storage_connection(
  tenant_id = "16decddb-28ac-4bea-8fc9-5844aadea669",
  app_id = "04b07795-8ddb-461a-bbee-02f9e1bf7b46",
  resource_endpoint = "https://eridev.dfs.core.windows.net/",
  storage_name = "projects",
  auth = "authorization_code",
  creds_yaml_path = NULL,
  ...
)
}
\arguments{
\item{tenant_id}{\code{str} ID of the Azure tenant}

\item{app_id}{\code{str} Application ID defaults to \code{"04b07795-8ddb-461a-bbee-02f9e1bf7b46"},
this can be changed if you have a service principal.}

\item{resource_endpoint}{\code{str} the URL used to connect to the Azure resource}

\item{storage_name}{\code{str} the name of the storage blob}

\item{auth}{\code{str} Authorization type defaults to \code{"authorization_code"},
this can be changed if you have a service principal.

Valid values are:\code{"authorization_code"}, \code{"device_code"},
\code{"client_credentials"}, \code{"resource_owner"}, \code{"on_behalf_of"}.

See \strong{Details} of \code{\link[AzureAuth:get_azure_token]{AzureAuth::get_azure_token()}} for further details.}

\item{creds_yaml_path}{\code{str} Path to the YAML file in Posit workbench.
If \code{NULL}, the path will be in \code{"~/credentials/posit_workbench_creds.yaml"}.}

\item{...}{additional parameters passed to \code{\link[AzureAuth:get_azure_token]{AzureAuth::get_azure_token()}}.}
}
\value{
Azure container verification
}
\description{
Generate token which connects to TCC Azure resources and
validates that the individual still has access. The current tenant ID
is hard coded for TCC RBLF resources.
}
\examples{
\dontrun{
azcontainer <- get_azure_storage_connection()
}

}
