% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{get_azure_storage_connection}
\alias{get_azure_storage_connection}
\title{Validate connection to Azure}
\usage{
get_azure_storage_connection(
  tenant_id = Sys.getenv("ERIFUNCTIONS_TENANT_ID"),
  app_id = Sys.getenv("ERIFUNCTIONS_APP_ID"),
  resource_endpoint = Sys.getenv("ERIFUNCTIONS_RESOURCE_ENDPOINT"),
  storage_name = Sys.getenv("ERIFUNCTIONS_STORAGE_NAME"),
  auth = "authorization_code",
  creds_yaml_path = NULL,
  ...
)
}
\arguments{
\item{tenant_id}{\code{str} ID of the Azure tenant. Defaults to \code{Sys.getenv("ERIFUNCTIONS_TENANT_ID")}.}

\item{app_id}{\code{str} Application ID. Defaults to \code{Sys.getenv("ERIFUNCTIONS_APP_ID")}.}

\item{resource_endpoint}{\code{str} URL used to connect to the Azure resource.
Defaults to \code{Sys.getenv("ERIFUNCTIONS_RESOURCE_ENDPOINT")}.}

\item{storage_name}{\code{str} Name of the storage blob.
Defaults to \code{Sys.getenv("ERIFUNCTIONS_STORAGE_NAME")}.}

\item{auth}{\code{str} Authorization type defaults to \code{"authorization_code"},
this can be changed if you have a service principal.

Valid values are:\code{"authorization_code"}, \code{"device_code"},
\code{"client_credentials"}, \code{"resource_owner"}, \code{"on_behalf_of"}.

See \strong{Details} of \code{\link[AzureAuth:get_azure_token]{AzureAuth::get_azure_token()}} for further details.}

\item{creds_yaml_path}{\code{str} Path to a YAML credentials file containing service principal
credentials. If \code{NULL} (default), interactive authentication is used via \code{auth}.}

\item{...}{additional parameters passed to \code{\link[AzureAuth:get_azure_token]{AzureAuth::get_azure_token()}}.}
}
\value{
Azure container object
}
\description{
Generate token which connects to TCC Azure resources and
validates that the individual still has access.
}
\examples{
\dontrun{
azcontainer <- get_azure_storage_connection()
}

}
