% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odk.R
\name{init_odk_connection}
\alias{init_odk_connection}
\title{Initialize/test ODK connection}
\usage{
init_odk_connection(
  url = Sys.getenv("ODK_URL", unset = "https://rblf.tccodk.org/"),
  user = Sys.getenv("ODK_USER", unset = ""),
  pass = Sys.getenv("ODK_PASS", unset = ""),
  testing = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{url}{\code{chr} Target URL for the ODK Central API}

\item{user}{\code{chr} Username}

\item{pass}{\code{chr} Password}

\item{testing}{\code{bool} Should results be evaluated for testing}

\item{verbose}{\code{bool} Should the function return character with information
or a boolean vector verifying connection}
}
\value{
\code{chr}/\code{bool} Verbose response or a T/F
}
\description{
Have a user prompt to enter username and password and then
save the resulting token to be used for this session
}
