% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odk.R
\name{list_odk_projects}
\alias{list_odk_projects}
\title{List ODK projects}
\usage{
list_odk_projects(
  url = Sys.getenv("ODK_URL"),
  auth = Sys.getenv("ODK_TOKEN"),
  testing = FALSE
)
}
\arguments{
\item{url}{\code{chr} Target URL for the ODK Central API}

\item{auth}{\code{chr} Authorization token to access URL}

\item{testing}{\code{bool} T/F if you want to just verify that the API if working}
}
\value{
\code{tibble} Output of all projects from API call
}
\description{
Given verified access, list all projects the user can access
}
