% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odk.R
\name{update_odk_app_user_role}
\alias{update_odk_app_user_role}
\title{Create/Delete/Assign/Un-assign app users}
\usage{
update_odk_app_user_role(
  action,
  project_id,
  form_id = NULL,
  actor_name = NULL,
  role_id = NULL,
  actor_id = NULL,
  testing = FALSE,
  url = Sys.getenv("ODK_URL"),
  auth = Sys.getenv("ODK_TOKEN")
)
}
\arguments{
\item{action}{\code{chr} 'create': Create a general app-user;
'delete': Delete an app-user;
'assign': Assign an app-user a role within a specific form
'revoke': Revoke an app-user role within a specific form
assign, revoke}

\item{project_id}{\code{int} The project id for which you want to identify users}

\item{form_id}{\code{chr} The form id for which you want to identify users}

\item{actor_name}{\code{chr} The display name of the actor to be updated}

\item{role_id}{\code{int} The role id which you want to assign, usually 2}

\item{actor_id}{\code{int} The actor id to be deleted}

\item{testing}{\code{bool} T/F if you want to just verify that the API if working}

\item{url}{\code{chr} Target URL for the ODK Central API}

\item{auth}{\code{chr} Authorization token to access URL}
}
\description{
Given an action, project id, form id, role id and actor id,
update the access that a specific app user is provided
}
